Geo/SearchPointVector: promote ThinToSize() tolerance factor to "double"
The unsigned integer will overflow very quickly after 65536, which is
not enough tolerance for some very deformed polygons, leading to an
endless loop and XCSoar freezing.

+++ src/Geo/SearchPointVector.cpp
@@ -37,7 +37,7 @@ bool
 SearchPointVector::ThinToSize(const unsigned max_size)
 {
   static constexpr double tolerance = 1.0e-8;
-  unsigned i = 2;
+  double i = 2;
   bool retval = false;
   while (size() > max_size) {
     retval |= ::PruneInterior(*this, tolerance * i);
